# ezkey-demo-app-acme/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ezkey Admin API",
    "description": "Administration API for Ezkey - Open Source Cryptographic MFA Platform\n\nThis API enables administrative management of Ezkey's main entities:\n- **Integrations**: Applications or systems protected by MFA\n- **Enrollments**: Associations between users, devices and integrations\n- **Auth Attempts**: MFA authentication attempts\n- **Admin Management**: Administrator authentication,\n                        enrollment recovery, and admin operations\n\nEzkey is intentionally distinct from FIDO2/WebAuthn and follows its own\ncryptographic MFA model. The API uses REST conventions and DTOs for all\nrequests and responses.\n",
    "contact": {
      "name": "Ezkey Team",
      "url": "https://ezkey.org",
      "email": "contributors@ezkey.org"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9080",
      "description": "Development server Admin API"
    },
    {
      "url": "https://admin-api.ezkey.org",
      "description": "Production server Admin API"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/enrollments/{id}/deactivate": {
      "post": {
        "tags": [
          "Enrollments"
        ],
        "summary": "Deactivate an enrollment",
        "description": "Reversibly deactivates an enrollment (active=false). Status remains VERIFIED. Can be undone with /reactivate. If the enrollment belongs to an admin, their active bearer tokens are immediately invalidated. Self-deactivation is not allowed.",
        "operationId": "deactivate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Enrollment ID to deactivate",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          },
          {
            "name": "reason",
            "in": "query",
            "description": "Optional justification for deactivation",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 500,
              "minLength": 0
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Enrollment deactivated successfully"
          },
          "400": {
            "description": "Invalid parameters or state conflict"
          },
          "403": {
            "description": "Access denied or self-deactivation"
          },
          "404": {
            "description": "Enrollment not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/auth-attempts": {
      "get": {
        "tags": [
          "Auth Attempts"
        ],
        "summary": "Search auth attempts",
        "description": "Retrieves authentication attempts with optional filters and pagination for security monitoring, forensic analysis, and compliance reporting. Supports dynamic sorting via ?sort=field,direction (e.g., ?sort=authAttemptId,asc). Default sort is by creation date descending (newest first).",
        "operationId": "search_2",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by authentication attempt status (PENDING, READ, ACCEPTED, REJECTED, INVALID, EXPIRED)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "READ",
                "INVALID",
                "REJECTED",
                "ACCEPTED",
                "EXPIRED"
              ]
            }
          },
          {
            "name": "enrollmentId",
            "in": "query",
            "description": "Filter by enrollment ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "integrationId",
            "in": "query",
            "description": "Filter by integration ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "createdAfter",
            "in": "query",
            "description": "Filter attempts created after this timestamp (ISO-8601)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "createdBefore",
            "in": "query",
            "description": "Filter attempts created before this timestamp (ISO-8601)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "default": [
                "createdAt,DESC"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search completed successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelAuthAttemptDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelAuthAttemptDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelAuthAttemptDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Auth Attempts"
        ],
        "summary": "Create new auth attempt",
        "description": "Creates a new authentication attempt for MFA validation",
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthAttemptCreateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Auth attempt created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthAttemptCreateRequestDto": {
        "type": "object",
        "description": "Auth attempt creation data",
        "properties": {
          "enrollmentId": {
            "type": "integer",
            "format": "int32",
            "description": "Enrollment ID (use this OR userIdentifier). Direct reference to enrollment.",
            "example": 123
          },
          "userIdentifier": {
            "type": "string",
            "description": "User identifier (username, user_id) for lookup within integration scope. Use this OR enrollmentId. When used with admin token, integrationId is required.",
            "example": "alice"
          },
          "integrationId": {
            "type": "integer",
            "format": "int32",
            "description": "Integration ID. Required when userIdentifier is used with admin token. Ignored when API key (derived from credentials).",
            "example": 1
          },
          "challengeRequested": {
            "type": "boolean",
            "description": "Whether a challenge code is requested for this attempt",
            "example": false
          },
          "contextTitle": {
            "type": "string",
            "description": "Optional short title for the approval request, displayed as the mobile card header. Max 200 characters. Example: \"Payment Approval\"",
            "example": "Payment Approval",
            "maxLength": 200,
            "minLength": 0
          },
          "contextMessage": {
            "type": "string",
            "description": "Optional descriptive message explaining what the approver is authorizing. Max 2000 characters.",
            "example": "Authorize payment batch #1497 to Acme Corp for $1,400",
            "maxLength": 2000,
            "minLength": 0
          },
          "demoMitmSignatureRequested": {
            "type": "boolean",
            "description": "Demo only: when true, this attempt is flagged for simulated MITM (tampered Pending body after signing) if Auth API ezkey.demo.mitm-signature-enabled is true.",
            "example": false
          }
        },
        "required": [
          "challengeRequested"
        ]
      },
      "AuthAttemptDto": {
        "type": "object",
        "description": "Authentication attempt details",
        "properties": {
          "authAttemptId": {
            "type": "integer",
            "format": "int32",
            "description": "Unique auth attempt ID",
            "example": 42
          },
          "enrollmentId": {
            "type": "integer",
            "format": "int32",
            "description": "Associated enrollment ID",
            "example": 7
          },
          "authAttemptStatus": {
            "type": "string",
            "description": "Current status of the authentication attempt",
            "enum": [
              "PENDING",
              "READ",
              "INVALID",
              "REJECTED",
              "ACCEPTED",
              "EXPIRED"
            ],
            "example": "PENDING"
          },
          "authAttemptChallenge": {
            "type": "integer",
            "format": "int32",
            "description": "Numeric challenge displayed to user for verification (null if not requested)",
            "example": 123456
          },
          "authAttemptProofToken": {
            "type": "string",
            "description": "Signed JWT proof token (present only when status is ACCEPTED)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp with timezone",
            "example": "2025-01-27T10:00:00+01:00"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Expiration timestamp with timezone",
            "example": "2025-01-27T10:05:00+01:00"
          },
          "contextTitle": {
            "type": "string",
            "description": "Short title for the approval request (null if no context provided)",
            "example": "Payment Approval"
          },
          "contextMessage": {
            "type": "string",
            "description": "Descriptive approval message (null if no context provided)",
            "example": "Authorize payment batch #1497 to Acme Corp for $1,400"
          },
          "demoMitmSignatureEnabled": {
            "type": "boolean",
            "description": "Demo MITM opt-in at creation time (Pending tampering only when Auth API demo flag is on)",
            "example": false
          }
        },
        "required": [
          "authAttemptId",
          "authAttemptStatus",
          "createdAt",
          "enrollmentId",
          "expiresAt"
        ]
      },
...
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PagedModelAuthAttemptDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthAttemptDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT authentication token for admin API",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}